+++ src/pages/selectJourneyDirection.tsx
@@ -1,21 +1,15 @@
 import React, { ReactElement } from 'react';
 import { NextPageContext } from 'next';
 import { parseCookies } from 'nookies';
-import flatMap from 'array.prototype.flatmap';
 import Layout from '../layout/Layout';
 import { OPERATOR_COOKIE, SERVICE_COOKIE, JOURNEY_COOKIE } from '../constants';
 import { deleteCookieOnServerSide } from '../utils';
-import {
-    getServiceByNocCodeAndLineName,
-    Service,
-    RawService,
-    RawJourneyPattern,
-    JourneyPattern, batchGetStopsByAtcoCode,
-} from '../data/auroradb';
+import { getServiceByNocCodeAndLineName, Service, RawService } from '../data/auroradb';
 import DirectionDropdown from '../components/DirectionDropdown';
 import FormElementWrapper from '../components/FormElementWrapper';
 import ErrorSummary from '../components/ErrorSummary';
 import { ErrorInfo } from '../types';
+import { enrichJourneyPatternsWithNaptanInfo } from '../utils/dataTransform';
 
 const title = 'Journey Selection - Fares data build tool';
 const description = 'Inbound and Outbound journey selection page of the Fares data build tool';
@@ -30,7 +24,7 @@ interface DirectionProps {
     inboundJourney: string;
 }
 
-const SelectJourneyDirection = ({ service, errors, outboundJourney, inboundJourney  }: DirectionProps): ReactElement => {
+const SelectJourneyDirection = ({ service, errors, outboundJourney, inboundJourney }: DirectionProps): ReactElement => {
     console.log('erros', errors);
     return (
         <Layout title={title} description={description}>
@@ -94,39 +88,6 @@ const SelectJourneyDirection = ({ service, errors, outboundJourney, inboundJourn
     );
 };
 
-const enrichJourneyPatternsWithNaptanInfo = async (journeyPatterns: RawJourneyPattern[]): Promise<JourneyPattern[]> =>
-    Promise.all(
-        journeyPatterns.map(
-            async (item: RawJourneyPattern): Promise<JourneyPattern> => {
-                const stopList = flatMap(item.JourneyPattern, stop => {
-                    return stop.OrderedStopPoints.map(stopPoint => stopPoint.StopPointRef);
-                });
-
-                const startPoint = item.JourneyPattern[0].OrderedStopPoints[0];
-                const [startPointStopLocality] = await batchGetStopsByAtcoCode([startPoint.StopPointRef]);
-
-                const endPoint = item.JourneyPattern.slice(-1)[0].OrderedStopPoints.slice(-1)[0];
-                const [endPointStopLocality] = await batchGetStopsByAtcoCode([endPoint.StopPointRef]);
-
-                return {
-                    startPoint: {
-                        Display: `${startPoint.CommonName}${
-                            startPointStopLocality?.localityName ? `, ${startPointStopLocality.localityName}` : ''
-                        }`,
-                        Id: startPoint.StopPointRef,
-                    },
-                    endPoint: {
-                        Display: `${endPoint.CommonName}${
-                            endPointStopLocality?.localityName ? `, ${endPointStopLocality.localityName}` : ''
-                        }`,
-                        Id: endPoint.StopPointRef,
-                    },
-                    stopList,
-                };
-            },
-        ),
-    );
-
 export const getServerSideProps = async (ctx: NextPageContext): Promise<{}> => {
     deleteCookieOnServerSide(ctx, JOURNEY_COOKIE);
     const cookies = parseCookies(ctx);
@@ -166,13 +127,13 @@ export const getServerSideProps = async (ctx: NextPageContext): Promise<{}> => {
     if (journeyCookie) {
         const journeyCookieInfo = JSON.parse(journeyCookie);
         if (journeyCookieInfo.errorMessages.length > 0) {
-            const { outboundJourney = '',  inboundJourney = '' } = journeyCookieInfo;
+            const { outboundJourney = '', inboundJourney = '' } = journeyCookieInfo;
             return {
                 props: {
                     service,
                     errors: journeyCookieInfo.errorMessages,
                     outboundJourney,
-                    inboundJourney
+                    inboundJourney,
                 },
             };
         }
